Solve the bug and separate closest test in its own file
