+++ package.json
@@ -31,9 +31,9 @@
     "eslint-plugin-compat": "^2.5.1",
     "eslint-plugin-import": "^2.14.0",
     "phantomjs-prebuilt": "^2.1.16",
-    "scriptkit": "0.4.0",
+    "scriptkit": "^1.0.0",
     "static-server": "^2.2.1",
-    "tap-spec": "^4.1.1",
+    "tap-spec": "^5.0.0",
     "tape-rollup": "4.6.4",
     "tape-run": "^2.1.5"
   },
